requirements.txt
Django==2.2.0
djangorestframework==3.10.0
flake8==3.7.0
djangorestframework-simplejwt
psycopg2-binary==2.8.5
Pillow

